Scaffold a host driver and CI jobs
Rather than bringing imxrt-hal and all the boards to this driver
package, we can bring this driver package to imxrt-hal and the boards.
I have a corresponding commit in imxrt-hal that stubs a USB host example
for the Teensy 4 and 1170EVK.

If we take that approach, this repository only has the driver, as a
library. We can still develop here and easily test the driver in the
imxrt-hal package, so I shared some quick docs on how to do that.